Calculate a timeseries of ETo across a set of input rasters.

Assumes that all inputs share the same resolution, extent, projection. Also assumes that rasters have the number of layers with each layer corresponding to a day. For hargreaves, only tmean, tmin and tmax are required arguments. For PM, tmean, srad, rh, and ws must all be supplied.

Usage

calc_etr_spatial(
  tmean,
  tmin = NULL,
  tmax = NULL,
  srad = NULL,
  rh = NULL,
  ws = NULL,
  elev = NULL,
  days = NULL,
  reference = 0.23,
  z = 9,
  method = "penman"
)

Arguments

tmean

A multilayer timeseries terra::rast of mean daily temperature in degrees C.

tmin

A multilayer timeseries terra::rast of max daily temperature in degrees C.

tmax

A multilayer timeseries terra::rast of min daily temperature in degrees C.

srad

A multilayer timeseries terra::rast of daily solar radiation in W m^-2.

rh

A multilayer timeseries terra::rast of mean daily relative humidity (%).

ws

A multilayer timeseries terra::rast of mean daily wind speed at 2m height in m s^-1.

elev

A terra::rast of elevation in meters. If left blank, elevation will be derived using the elevatr package.

days

A multilayer timeseries terra::rast where each day is a constant value of the Julian day. If left blank, it is assumed the tmean input has a time attribute and the raster will be derived using terra::time.

reference

The albedo of the reference surface ranging from 0 - 1. Defaults to 0.23 for grass.

z

The zoom level download elevation data at ranging from 1 - 14. One if coarser resolution, 14 is finer resolution. For more information, look at ?elevatr::get_elev_raster

method

The method to calculate ETo. Can either be "penman" or "hargreaves".

Value

A terra::rast timeseries of ETo for the input domain.

Examples

if (FALSE) {
# Load data. Need to read with terra::rast to unpack to a raster.
srad <- terra::rast(srad) |> terra::subset(1:10)
tmean <- terra::rast(tmean) |> terra::subset(1:10)
# Convert from K to C
tmean <- tmean - 273.15
tmax <- terra::rast(tmax) |> terra::subset(1:10)
# Convert from K to C
tmax <- tmax - 273.15
tmin <- terra::rast(tmin) |> terra::subset(1:10)
# Convert from K to C
tmin <- tmin - 273.15
rh <- terra::rast(rh) |> terra::subset(1:10)
ws <- terra::rast(ws) |> terra::subset(1:10)

penman <- calc_etr_spatial(
  tmean = tmean, srad = srad, rh = rh, ws = ws,
  method = "penman", reference = 0.23, z = 3
)
hargreaves <- calc_etr_spatial(
  tmean = tmean, tmax = tmax, tmin = tmin, method = "hargreaves", z = 3
)
}